refactor(walletconnect): split multichain connector into per-chain adapters
