+++ app/core/WalletConnect/wc-utils.ts
@@ -30,11 +30,6 @@ import { TrxAccountType, TrxScope } from '@metamask/keyring-api';
 ///: END:ONLY_INCLUDE_IF
 import Engine from '../Engine';
 import { APPROVED_METHODS_BY_NAMESPACE } from './wc-config';
-import {
-  getCompatibleTronCaipChainIdsForWalletConnect,
-  normalizeCaipChainIdInboundForWalletConnect,
-  normalizeCaipChainIdOutboundForWalletConnect,
-} from './WalletConnectMultiChainConnector';
 
 export interface WCMultiVersionParams {
   protocol: string;
@@ -264,6 +259,65 @@ export const normalizeCaipChainIdOutbound = (caipChainId: string): string => {
   const normalized = normalizeCaipChainIdOutboundForWalletConnect(caipChainId);
   DevLogger.log(`WC::normalizeCaipChainIdOutbound normalized=${normalized}`);
   return normalized;
+};
+
+export function normalizeCaipChainIdInboundForWalletConnect(
+  caipChainId: string,
+): string {
+  const namespace = caipChainId.split(':')[0];
+
+  if (namespace === KnownCaipNamespace.Tron) {
+    const chainRef = caipChainId.slice('tron:'.length);
+    if (chainRef.startsWith('0x')) {
+      const decimalRef = parseInt(chainRef, 16);
+      const normalized = `tron:${decimalRef}`;
+      DevLogger.log('[wc][caip] normalize inbound tron chainId hex->dec', {
+        input: caipChainId,
+        output: normalized,
+      });
+      return normalized;
+    }
+  }
+
+  return caipChainId;
+}
+
+export function normalizeCaipChainIdOutboundForWalletConnect(
+  caipChainId: string,
+): string {
+  const namespace = caipChainId.split(':')[0];
+
+  if (namespace === KnownCaipNamespace.Tron) {
+    const chainRef = caipChainId.slice('tron:'.length);
+    if (!chainRef.startsWith('0x')) {
+      const hexRef = parseInt(chainRef, 10).toString(16);
+      const normalized = `tron:0x${hexRef}`;
+      DevLogger.log('[wc][caip] normalize outbound tron chainId dec->hex', {
+        input: caipChainId,
+        output: normalized,
+      });
+      return normalized;
+    }
+  }
+
+  return caipChainId;
+}
+
+export const getCompatibleTronCaipChainIdsForWalletConnect = (
+  caipChainId: string,
+): string[] => {
+  if (!caipChainId.startsWith(`${KnownCaipNamespace.Tron}:`)) {
+    return [caipChainId];
+  }
+
+  const inboundNormalized =
+    normalizeCaipChainIdInboundForWalletConnect(caipChainId);
+  const outboundNormalized =
+    normalizeCaipChainIdOutboundForWalletConnect(caipChainId);
+
+  return Array.from(
+    new Set([caipChainId, inboundNormalized, outboundNormalized]),
+  );
 };
 
 export const getScopedPermissions = async ({
